review comment

+++ src/Storages/MergeTree/ReplicatedMergeTreeSink.cpp
@@ -356,7 +356,6 @@ void ReplicatedMergeTreeSink::consume(Chunk & chunk)
         if (total_streams + current_streams > max_insert_delayed_streams_for_parallel_write)
         {
             finishDelayed(zookeeper);
-            //std::move(current_parts.begin(), current_parts.end(), std::back_inserter(delayed_parts));
             delayed_parts = std::move(current_parts);
             finishDelayed(zookeeper);
 
@@ -414,7 +413,7 @@ void ReplicatedMergeTreeSink::finishDelayed(const ZooKeeperWithFaultInjectionPtr
             /// reset the cache version to zero for every partition write.
             /// Version zero allows to avoid wait on first iteration
             cache_version = 0;
-            int retry_times = 0;
+            size_t retry_times = 0;
             while (true)
             {
                 partition.temp_part->finalize();
